+++ backend/cloudtasks/appengine/http_target_tasks_handler.go
@@ -15,7 +15,7 @@ func (h *Handlers) HttpTargetTasksHandler(w http.ResponseWriter, r *http.Request
 	log.InfoKV(ctx, "HttpTargetTasksHandler.request.header", r.Header)
 
 	// Cloud Tasks Http Target Task を IAP 貫通させると X-Goog-Iap-Jwt-Assertion が付いてる
-	if err := ValidateJWTFromAppEngine(r, h.projectNumber, h.projectID); err != nil {
+	if err := ValidateJWTFromAppEngine(ctx, r, h.projectNumber, h.projectID); err != nil {
 		aelog.Errorf(ctx, "failed ValidateJWTFromAppEngine. pn:%s,pID:%s, %v\n", h.projectNumber, h.projectID, err)
 	}
