+++ go.mod
@@ -4,11 +4,10 @@ go 1.14
 
 require (
 	cloud.google.com/go v0.68.0
-	github.com/k0kubun/pp v3.0.1+incompatible
-	github.com/mattn/go-colorable v0.1.8 // indirect
 	github.com/sinmetalcraft/gcpbox v1.0.0
 	github.com/vvakame/sdlog v0.0.0-20200409072131-7c0d359efddc
 	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1
 	google.golang.org/api v0.32.0
 	google.golang.org/genproto v0.0.0-20201002142447-3860012362da
+	google.golang.org/protobuf v1.25.0
 )
